Update Experiment Server to support Comet Hunters / VOLCROWE experiment
This is a combination of 18 commits.

Added Comet hunters experiment
Added tests: basic and end-to-end
Allow use of a different project when on staging
Test for all cohorts.
Update the end of experiment logic so that we never become inactive.
Ensure that intervention_time is always correct.
Improved error handling when tests fail due to network issues
Improve tests: "We're running out of elements!"
Add an extra sanity check to ensure we never run out of IDs again in the test.